Windows: establish Redis 7.2 integration checkpoint
Preserve the first buildable and restart-valid direct 7.2.14 merge resolution across MinGW static runtime linkage, FDAPI/IOCP, QFork/jemalloc, multi-part AOF, replication handoff, modules, and the rebased Windows test harness.

+++ src/acl.c
@@ -1510,10 +1510,10 @@ unsigned long ACLGetCommandID(sds cmdname) {
     void *id = raxFind(commandId,(unsigned char*)lowername,sdslen(lowername));
     if (id != raxNotFound) {
         sdsfree(lowername);
-        return (unsigned long)id;
+        return (unsigned long)(uintptr_t)id;
     }
     raxInsert(commandId,(unsigned char*)lowername,strlen(lowername),
-              (void*)nextid,NULL);
+              (void*)(uintptr_t)nextid,NULL);
     sdsfree(lowername);
     unsigned long thisid = nextid;
     nextid++;
