+++ tests/modules/blockonbackground.c
@@ -72,8 +72,8 @@ void HelloBlock_FreeStringData(RedisModuleCtx *ctx, void *privdata) {
 void *BlockDebug_ThreadMain(void *arg) {
     void **targ = arg;
     RedisModuleBlockedClient *bc = targ[0];
-    long long delay = (unsigned long)targ[1];
-    long long enable_time_track = (unsigned long)targ[2];
+    long long delay = (long long)(intptr_t)targ[1];
+    long long enable_time_track = (long long)(intptr_t)targ[2];
     BlockPrivdata *block_privdata = RedisModule_BlockClientGetPrivateData(bc);
 
     if (enable_time_track)
@@ -96,7 +96,7 @@ void *BlockDebug_ThreadMain(void *arg) {
 void *DoubleBlock_ThreadMain(void *arg) {
     void **targ = arg;
     RedisModuleBlockedClient *bc = targ[0];
-    long long delay = (unsigned long)targ[1];
+    long long delay = (long long)(intptr_t)targ[1];
     BlockPrivdata *block_privdata = RedisModule_BlockClientGetPrivateData(bc);
     blockClientMeasureTimeStart(bc, block_privdata);
     RedisModule_Free(targ);
@@ -151,7 +151,7 @@ int HelloBlock_RedisCommand(RedisModuleCtx *ctx, RedisModuleString **argv, int a
      * the delay and a reference to the blocked client handle. */
     void **targ = RedisModule_Alloc(sizeof(void*)*3);
     targ[0] = bc;
-    targ[1] = (void*)(unsigned long) delay;
+    targ[1] = (void*)(intptr_t)delay;
     // pass 1 as flag to enable time tracking
     targ[2] = (void*)(unsigned long) 1;
 
@@ -193,7 +193,7 @@ int HelloBlockNoTracking_RedisCommand(RedisModuleCtx *ctx, RedisModuleString **a
      * the delay and a reference to the blocked client handle. */
     void **targ = RedisModule_Alloc(sizeof(void*)*3);
     targ[0] = bc;
-    targ[1] = (void*)(unsigned long) delay;
+    targ[1] = (void*)(intptr_t)delay;
     // pass 0 as flag to enable time tracking
     targ[2] = (void*)(unsigned long) 0;
 
@@ -225,7 +225,7 @@ int HelloDoubleBlock_RedisCommand(RedisModuleCtx *ctx, RedisModuleString **argv,
      * the delay and a reference to the blocked client handle. */
     void **targ = RedisModule_Alloc(sizeof(void*)*2);
     targ[0] = bc;
-    targ[1] = (void*)(unsigned long) delay;
+    targ[1] = (void*)(intptr_t)delay;
 
     if (pthread_create(&tid,NULL,DoubleBlock_ThreadMain,targ) != 0) {
         RedisModule_AbortBlock(bc);
